LayoutTests/platform/gtk/accessibility/document-reload-events.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script>

function runTest()
{
  if (!layoutTestController.globalFlag) {
    layoutTestController.globalFlag = true;

    if (window.layoutTestController) {
      layoutTestController.dumpAsText();
      layoutTestController.waitUntilDone();
    }

    setTimeout("window.location.reload()", 0);
  } else {
    layoutTestController.globalFlag = false;
    if (window.layoutTestController)
      layoutTestController.notifyDone();
  }
}
</script>
</head>
<body id="body" onload="runTest();">
<p id="description"></p>
<div id="console"></div>
<script>
description("Test that AtkDocument emits the right signals when reloading.");

if (window.accessibilityController)
  accessibilityController.logAccessibilityEvents();

</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
