safari-534.53-branch/LayoutTests/platform/gtk/accessibility/document-reload-events-expected.txt

Accessibility object emitted "load-complete" / Name: "(No name)" / Role: 80
Accessibility object emitted "state-change:busy = 0" / Name: "(No name)" / Role: 80
Accessibility object emitted "state-change:busy = 1" / Name: "(No name)" / Role: 80
Accessibility object emitted "reload" / Name: "(No name)" / Role: 80
Accessibility object emitted "state-change:defunct = 1" / Name: "(No name)" / Role: 80
Accessibility object emitted "load-complete" / Name: "(No name)" / Role: 80
Accessibility object emitted "state-change:busy = 0" / Name: "(No name)" / Role: 80
Test that AtkDocument emits the right signals when reloading.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
